LayoutTests/fast/events/keydown-keypress-preventDefault.html

<html>
<script>
    function test() {
        var tf = document.getElementById('tf');
        tf.focus();
        if (window.layoutTestController) {
            layoutTestController.dumpAsText();
            eventSender.keyDown('F');
            eventSender.keyDown('A');
            eventSender.keyDown('I');
            eventSender.keyDown('L');
        }
    }
    
    function log(msg) {
        var res = document.getElementById('res');
        res.innerHTML = res.innerHTML + msg + "<br>";
    }
</script> 
<body onload="test()">
    This tests that preventing the default behavior for a keydown event will also prevent the keypress event from firing, and the text from being inserted.<br>
    This matches IE behavior, but not Firefox.<br>
    <input id="tf" onkeydown="log('key down'); event.preventDefault();" onkeypress="log('key press')">
    <br>
    <div id="res"></div>
</body>
</html>
