LayoutTests/fast/events/keydown-keypress-preventDefault-expected.txt

This tests that preventing the default behavior for a keydown event will also prevent the keypress event from firing, and the text from being inserted.
This matches IE behavior, but not Firefox.
 
key down
key down
key down
key down
